main to access modules
With these change we can see the main.tf that will use the modules created earlier.
One of the weakness of the origin code was that it used specific values which goes against the reusability and flexibility of the terraform modules. in order to get that flexibility theses values are brought back here.
In order to use the variables in the main.tf, I also changed the variables.tf accordingly, and the same goes for the outputs.tf